Finish initial implementation

+++ models/Inventory.js
@@ -7,5 +7,5 @@ module.exports = (sequelize, DataTypes) => sequelize.define('inventory', {
   },
   name: DataTypes.STRING, // Item name
   description: DataTypes.STRING, // Item description
-  quantity: DataTypes.INTEGER, // Quantity of items in stock
+  quantity: DataTypes.FLOAT, // Quantity of item in stock
 });
